Adjustment: Type of minutes

+++ BackEnd/routers/flashcards.py
@@ -12,7 +12,7 @@ class FlashcardSessionIn(SQLModel):
     deck_id: int
     correct: int
     total: int
-    elapsed_minutes: int  # ⬅️ novo
+    elapsed_minutes: float  # ⬅️ novo
 
 @router.post("/session-result")
 def finish_session(
